requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.17
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.8.4
idna==2.10
mccabe==0.6.1
Pillow==8.1.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.0
pytz==2020.5
razorpay==1.2.0
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
